chore: header 내 현재 날짜 노출 영역 component 분리
