+++ client/components/common/Header.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import {useRouter} from 'next/router';
-
 import {
     useRecoilState,
     useRecoilValue,
@@ -16,39 +14,16 @@ import {
     viewState
 } from '../../recoil/atoms';
 
-import {
-    ViewType
-} from '../../utils/constants';
-
 import {useToggleModal} from '../../hooks/useCloseModal';
 
 import {CalendarDirection} from './CalendarDirection';
+import {CalendarHeading} from './CalendarHeading';
 import {Icon} from './Icons';
 import {ButtonText} from './ButtonText';
 
 export const HeaderComponent = () => {
     const [aside, setAside] = useRecoilState(asideState);
-
-    const view = useRecoilValue(viewState);
-    const {type} = view;
-
     const curr = useRecoilValue(targetState);
-    const {full, fullYear, month, date, day} = curr;
-
-    const [updateCurr, setUpdateCurr] = useRecoilState(targetStateState);
-
-    const setMonth = () => {
-        if (type === ViewType.Day || type === ViewType.Month) {
-            return Number(month) + 1;
-        }
-
-        if (Number(date) + (type === ViewType.Week ? 6 : 2) > updateCurr?.monthLastNumber) {
-            const calcYear = month === 11 ? `${Number(fullYear) + 1} / 1` : Number(month) + 2;
-            return `${Number(month) + 1} - ${calcYear}`;
-        }
-
-        return `${Number(month) + 1}`;
-    }
 
     return (
         <StyledHeader>
@@ -61,16 +36,8 @@ export const HeaderComponent = () => {
                 <Icon iconType="hamburger"/>
                 <ButtonText a11y={true}>보기 옵션 {aside.isVisible ? '닫기' : '열기'}</ButtonText>
             </StyledButton>
-            <CalendarDirection/>
-            <StyledHeading>
-                {full && <StyledDateWrap>
-                    <StyledDateElement>{Number(fullYear)}</StyledDateElement>
-                    {type !== ViewType.Year && <StyledDateElement>
-                        {setMonth()}
-                    </StyledDateElement>}
-                    {type === ViewType.Day && <StyledDateElement>{Number(date)}</StyledDateElement>}
-                </StyledDateWrap>}
-            </StyledHeading>
+            {curr && <CalendarDirection/>}
+            {curr && <CalendarHeading/>}
         </StyledHeader>
     );
 };
@@ -83,44 +50,6 @@ const StyledHeader = styled.header`
   padding: 8px 15px 7px;
   box-sizing: border-box;
   border-bottom: solid 1px var(--light-gray-color);
-`;
-
-const StyledHeading = styled.h1`
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const StyledDateWrap = styled.span`
-  display: inline-flex;
-  align-items: center;
-`;
-
-const StyledDateElement = styled.span`
-  display: inline-flex;
-  font-size: 26px;
-
-  + span {
-    &:before {
-      content: "/";
-      display: inline-flex;
-      position: relative;
-      margin: 0 4px;
-    }
-  }
-`;
-
-const StyledButtonWrap = styled.span`
-  display: flex;
-  align-items: center;
-  gap: 5px;
-
-  button {
-    &:first-child {
-      margin-right: 40px;
-    }
-  }
 `;
 
 const StyledButton = styled.button`
